requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
psycopg2-binary
tensorflow
joblib
kafka-python
python-dotenv
